Cleaned up the structure of Calendar.dart, created a directory for calendar files
